DLK-714: Adds audit provider metrics to play auditing
This is needed as part of a change to send audit counts to datastream and logs.

It separates the construction of audit events from the sending of audit events.

The AuditChannel code will be used by both the AuditConnector and the new AuditCounter

In order to identify audit loss audit counters are captured and emmited as logs and through the auditing channel

See https://confluence.tools.tax.service.gov.uk/display/TM/Auditing+Counters for more details about the motivations